update: 프론트 - 유저정보 미들웨어 추가 - 모든 페이지 템플릿 제작 - 모든 페이지 이동 연결
