+++ src/views/storeAdmin.ejs
@@ -30,28 +30,39 @@
         <div class="max-w-4xl mx-auto p-8">
             <div class="bg-white shadow-md rounded p-6">
                 <div class="flex flex-col lg:flex-row gap-6">
-                    <div class="image-placeholder w-full lg:w-1/2 h-64 rounded">이미지 파일</div>
+                    <div class="image-placeholder w-full lg:w-1/2 h-64 rounded">
+                        <img src="" id="storeImage" class="w-full h-full object-cover" />
+                    </div>
                     <div class="flex-1 space-y-4">
                         <div>
                             <label for="storeName" id="storeName" class="block text-sm font-medium text-gray-700">스토어명</label>
+                            <label for="storeAddress" id="storeAddress" class="mt-1 block w-full border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm">주소</label>
                             <!-- <input type="text" id="storeName" name="storeName" placeholder="스토어이름" class="mt-1 block w-full border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"> -->
                         </div>
                         <div class="grid grid-cols-3 gap-4">
-                            <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" type="button">사진바꾸기</button>
-                            <button class="bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" type="button">메뉴관리</button>
-                            <button class="bg-yellow-500 hover:bg-yellow-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" type="button">주문목록</button>
+                        <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" type="button">
+                        사진바꾸기
+                        </button>
+                        <button class="bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" type="button">
+                        메뉴관리
+                        </button>
+                        <button class="bg-yellow-500 hover:bg-yellow-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" type="button">
+                        주문목록
+                        </button>
                         </div>
                     </div>
                 </div>
                 <div class="mt-6">
-                    <button class="bg-indigo-500 hover:bg-indigo-700 text-white font-bold py-3 px-6 w-full rounded focus:outline-none focus:shadow-outline" type="button">스토어 수정</button>
+                <button class="bg-indigo-500 hover:bg-indigo-700 text-white font-bold py-3 px-6 w-full rounded focus:outline-none focus:shadow-outline" type="button">
+                스토어 수정
+                </button>
                 </div>
             </div>
 
             <div id="menu-container"></div>
         </div>
         <%-include('footer')%>
-        <div id="storeId" data-store-id="<%=storeId%>"></div>
+        <div id="storeId" data-store-id="<%=data.store.id%>"></div>
     </body>
-    <script src="<%= path %>javascripts/shop.js"></script>
+    <!-- <script src="<%= path %>javascripts/store.js"></script> -->
 </html>
